+++ Packages/eu.netherlands3d.selection-tools/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2025-10-21
+
+### Changed
+
+* The areaselection for grids now inherits from polygon input. 
+* Added an extra state to drawmode, "Selected" to indicate that the area is selected but not being edited.
+
+
 ## [3.2.3] - 2025-10-20
 
 ### Fixed